// src/model/dto/character.dto.ts
export interface CharacterDto {
  _id: string;
  name: string;
  role: string;
  house: string;
  school: string;
  __v: number;
  ministryOfMagic: boolean;
  orderOfThePhoenix: boolean;
  dumbledoresArmy: boolean;
  deathEater: boolean;
  bloodStatus: string;
  species: string;
  boggart: string;
  alias: string;
  animagus: string;
  wand: string;
  patronus: string;
}